@@ -1,76 +1,59 @@
 {
 	"name": "frontend-final",
 	"version": "0.0.0",
-	"engines": {
-		"node": "14.17.1",
-		"npm": "6.14.13"
-	},
 	"scripts": {
 		"ng": "ng",
 		"start": "node server.js",
 		"test-start": "ng build && node server.js",
 		"serve": "ng serve",
 		"build": "ng build",
-		"test": "ng test",
-		"lint": "ng lint",
-		"e2e": "ng e2e",
-		"lite": "lite-server",
 		"preinstall": "npm install -g @angular/cli @angular/compiler-cli typescript",
-		"postinstall": "",
-		"staging-build": "ng build --configuration=staging",
-		"proxy-start": "ng serve --proxy-config proxy.conf.json"
+		"postinstall": "ng build --prod"
 	},
 	"private": true,
 	"dependencies": {
-		"@angular/animations": "^12.2.16",
-		"@angular/cdk": "^13.3.2",
-		"@angular/common": "~12.1.2",
-		"@angular/compiler": "~12.1.2",
-		"@angular/core": "~12.1.2",
-		"@angular/forms": "~12.1.2",
-		"@angular/localize": "~12.1.2",
-		"@angular/material": "^13.3.2",
-		"@angular/platform-browser": "~12.1.2",
-		"@angular/platform-browser-dynamic": "~12.1.2",
-		"@angular/router": "~12.1.2",
+		"@angular/animations": "~13.3.0",
+		"@angular/common": "~13.3.0",
+		"@angular/compiler": "~13.3.0",
+		"@angular/core": "~13.3.0",
+		"@angular/forms": "~13.3.0",
+		"@angular/localize": "~13.3.0",
+		"@angular/platform-browser": "~13.3.0",
+		"@angular/platform-browser-dynamic": "~13.3.0",
+		"@angular/router": "~13.3.0",
 		"@auth0/angular-jwt": "^5.0.2",
 		"@fortawesome/angular-fontawesome": "^0.10.2",
-		"@fortawesome/fontawesome-svg-core": "^1.2.36",
-		"@fortawesome/free-brands-svg-icons": "^5.15.4",
-		"@fortawesome/free-regular-svg-icons": "^6.1.1",
-		"@fortawesome/free-solid-svg-icons": "^5.15.4",
-		"@ng-bootstrap/ng-bootstrap": "^10.0.0",
-		"@ngx-translate/core": "^14.0.0",
-		"@ngx-translate/http-loader": "^7.0.0",
-		"@ngxs/logger-plugin": "^3.7.3",
-		"@ngxs/storage-plugin": "^3.7.3",
+		"@fortawesome/fontawesome-svg-core": "^6.1.0",
+		"@fortawesome/free-brands-svg-icons": "^6.1.0",
+		"@fortawesome/free-regular-svg-icons": "^6.1.0",
+		"@fortawesome/free-solid-svg-icons": "^6.1.0",
+		"@ng-bootstrap/ng-bootstrap": "^12.0.2",
+		"@ngxs/logger-plugin": "^3.7.3-dev.master-33cace6",
 		"@ngxs/store": "^3.7.3",
-		"bootstrap": "^4.5.0",
-		"chart.js": "^3.7.1",
-		"cors": "^2.8.5",
-		"express": "^4.17.3",
+		"@popperjs/core": "^2.10.2",
+		"angular-animations": "^0.11.0",
+		"apexcharts": "^3.35.0",
+		"bootstrap": "^5.1.3",
 		"jwt-decode": "^3.1.2",
-		"ng-bootstrap-form-validation": "^9.0.1",
+		"ng-apexcharts": "^1.7.1",
 		"ngx-toastr": "^14.2.3",
-		"path": "^0.12.7",
-		"remove": "^0.1.5",
-		"rxjs": "~6.6.0",
-		"tslib": "^2.2.0",
-		"zone.js": "^0.11.5"
+		"rxjs": "~7.5.0",
+		"tslib": "^2.3.0",
+		"zone.js": "~0.11.4"
 	},
 	"devDependencies": {
-		"@angular-devkit/build-angular": "~12.1.2",
-		"@angular/cli": "~12.1.2",
-		"@angular/compiler-cli": "~12.1.2",
+		"@angular-devkit/build-angular": "~13.3.2",
+		"@angular/cli": "~13.3.2",
+		"@angular/compiler-cli": "~13.3.0",
 		"@ngxs/devtools-plugin": "^3.7.3",
-		"@types/jasmine": "~3.8.0",
-		"@types/node": "^12.20.47",
-		"jasmine-core": "~3.8.0",
-		"karma": "^6.3.17",
-		"karma-chrome-launcher": "^3.1.1",
-		"karma-coverage": "~2.0.3",
-		"karma-jasmine": "^4.0.2",
+		"@types/jasmine": "~3.10.0",
+		"@types/node": "^12.11.1",
+		"jasmine-core": "~4.0.0",
+		"karma": "~6.3.0",
+		"karma-chrome-launcher": "~3.1.0",
+		"karma-coverage": "~2.1.0",
+		"karma-jasmine": "~4.0.0",
 		"karma-jasmine-html-reporter": "~1.7.0",
-		"typescript": "~4.3.2"
+		"typescript": "~4.6.2"
 	}
 }
