@@ -1,39 +1,41 @@
-import { Component } from '@angular/core';
-import { Title } from '@angular/platform-browser';
-import { ChildActivationEnd, Router } from '@angular/router';
-import { TranslateService } from '@ngx-translate/core';
-import { Actions, ofActionDispatched, Store } from '@ngxs/store';
-import { Logout } from '@shared/auth/infrastructure/ngxs/auth.actions';
-import { filter } from 'rxjs/operators';
+import { Component, OnInit } from '@angular/core';
+import { Select, Store } from '@ngxs/store';
+import { Observable } from 'rxjs';
+import { Auth } from './auth/domain/auth.model';
+import { AuthService } from './auth/services/auth.service';
+import { Login, Logout } from './auth/state/auth.actions';
+import { AuthState } from './auth/state/auth.state';
 
 @Component({
-	selector: 'app-root',
-	template: `<router-outlet></router-outlet>
-	<app-loader></app-loader>
-	`,
+  selector: 'app-root',
+  template: `
+  <button (click)="logoutHandler()">Logout</button>
+  <router-outlet></router-outlet>
+  `
 })
-export class AppComponent {
-	private idiomas: string[];
-	title = 'AppComponent';
-	constructor(public router: Router, private titleService: Title, private store: Store, private actions: Actions, public translate: TranslateService) {
-		//lenguajes
-		this.idiomas = ['es-ES', 'en'];
-		translate.addLangs(this.idiomas);
-		translate.setDefaultLang('es-ES');
-		//title
-		this.router.events
-			.pipe(
-				filter(event => event instanceof ChildActivationEnd))
-			.subscribe(event => {
-				let snapshot = (event as ChildActivationEnd).snapshot;
-				while (snapshot.firstChild !== null) {
-					snapshot = snapshot.firstChild;
-				}
-				this.titleService.setTitle(snapshot.data.title || this.title);
-			});
-		//store
-		this.actions.pipe(ofActionDispatched(Logout)).subscribe(() => {
-			this.router.navigate(['/login']);
-		})
-	}
+export class AppComponent implements OnInit {
+  constructor(
+    private store: Store,
+    public authService: AuthService
+  ) { }
+
+  ngOnInit() {
+    /* const jwtToken = this.authService.checkToken();
+    // Replace this with real object
+    const userMokeado: Auth = {
+      id: 1,
+      username: "",
+      password: "",
+      token: jwtToken,
+      role: "customer"
+    }
+
+    if (jwtToken) {
+      this.store.dispatch(new Login(userMokeado));
+      this.authService.checkRole(jwtToken);
+    } */
+  }
+  logoutHandler() {
+    this.store.dispatch(new Logout(0));
+  }
 }
