@@ -1,5 +1,5 @@
 export const environment = {
-	production: true,
-	//baseUrl: 'http://localhost:8000'
-	baseUrl: 'https://back-final-xoje.herokuapp.com'
+  production: true,
+  baseUrl: 'https://back-final-xoje.herokuapp.com'
+
 };
