@@ -3,9 +3,8 @@
 // The list of file replacements can be found in `angular.json`.
 
 export const environment = {
-	production: false,
-	baseUrl: 'http://localhost:8000'
-	// baseUrl: 'https://back-final-xoje.herokuapp.com'
+  production: false,
+  baseUrl: 'http://localhost:8000/api'
 };
 
 /*
